feat: use progress-bar style to replay user recording for synthesis ref audio comparison
